+++ GBAlgCource.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31911.196
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GB_AlgCource", "MainProject\GB_AlgCource.csproj", "{63C4D65F-54A0-41F0-8996-2D968B226EB6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GBAlgCource", "GBAlgCource\GBAlgCource.csproj", "{63C4D65F-54A0-41F0-8996-2D968B226EB6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
