+++ GBAlgCource/Lesson1/Lesson1_2.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace GB_AlgCource
+namespace GBAlgCource
 {
     internal class Lesson2StrangeSum : ILesson
     {
