+++ GBAlgCource/Lesson1/Lesson1_3.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace GB_AlgCource
+namespace GBAlgCource
 {
     internal class Lesson3Fibonacci : ILesson
     {
